Reversion.

+++ Include/jsonifier/Compare.hpp
@@ -44,14 +44,14 @@ namespace jsonifier_internal {
 		static constexpr char value{ valueNewer };
 		JSONIFIER_FORCE_INLINE static const char_type* memchar(const char_type* data, size_t lengthNew) noexcept {
 #if JSONIFIER_CHECK_FOR_INSTRUCTION(JSONIFIER_AVX512)
-			if (static_cast<int64_t>(lengthNew) >= 64) {
-				using simd_type						= typename get_type_at_index<simd_internal::avx_list, 2>::type::type;
-				using integer_type					= typename get_type_at_index<simd_internal::avx_list, 2>::type::integer_type;
-				static constexpr int64_t vectorSize = get_type_at_index<simd_internal::avx_list, 2>::type::bytesProcessed;
-				const simd_type search_value		= simd_internal::gatherValue<simd_type>(static_cast<uint8_t>(value));
+			if (lengthNew >= 64) {
+				using simd_type					   = typename get_type_at_index<simd_internal::avx_list, 2>::type::type;
+				using integer_type				   = typename get_type_at_index<simd_internal::avx_list, 2>::type::integer_type;
+				static constexpr size_t vectorSize = get_type_at_index<simd_internal::avx_list, 2>::type::bytesProcessed;
+				const simd_type search_value	   = simd_internal::gatherValue<simd_type>(static_cast<uint8_t>(value));
 				integer_type mask;
 				simd_type chunk;
-				while (static_cast<int64_t>(lengthNew) >= vectorSize) {
+				while (lengthNew >= vectorSize) {
 					chunk = simd_internal::gatherValuesU<simd_type>(data);
 					mask  = simd_internal::opCmpEq(chunk, search_value);
 					if JSONIFIER_UNLIKELY (mask != 0) {
@@ -65,14 +65,14 @@ namespace jsonifier_internal {
 #endif
 
 #if JSONIFIER_CHECK_FOR_INSTRUCTION(JSONIFIER_AVX512) || JSONIFIER_CHECK_FOR_INSTRUCTION(JSONIFIER_AVX2)
-			if (static_cast<int64_t>(lengthNew) >= 32) {
-				using simd_type						= typename get_type_at_index<simd_internal::avx_list, 1>::type::type;
-				using integer_type					= typename get_type_at_index<simd_internal::avx_list, 1>::type::integer_type;
-				static constexpr int64_t vectorSize = get_type_at_index<simd_internal::avx_list, 1>::type::bytesProcessed;
-				const simd_type search_value		= simd_internal::gatherValue<simd_type>(static_cast<uint8_t>(value));
+			if (lengthNew >= 32) {
+				using simd_type					   = typename get_type_at_index<simd_internal::avx_list, 1>::type::type;
+				using integer_type				   = typename get_type_at_index<simd_internal::avx_list, 1>::type::integer_type;
+				static constexpr size_t vectorSize = get_type_at_index<simd_internal::avx_list, 1>::type::bytesProcessed;
+				const simd_type search_value	   = simd_internal::gatherValue<simd_type>(static_cast<uint8_t>(value));
 				integer_type mask;
 				simd_type chunk;
-				while (static_cast<int64_t>(lengthNew) >= vectorSize) {
+				while (lengthNew >= vectorSize) {
 					chunk = simd_internal::gatherValuesU<simd_type>(data);
 					mask  = simd_internal::opCmpEq(chunk, search_value);
 					if JSONIFIER_UNLIKELY (mask != 0) {
@@ -86,14 +86,14 @@ namespace jsonifier_internal {
 #endif
 #if JSONIFIER_CHECK_FOR_INSTRUCTION(JSONIFIER_AVX512) || JSONIFIER_CHECK_FOR_INSTRUCTION(JSONIFIER_AVX2) || JSONIFIER_CHECK_FOR_INSTRUCTION(JSONIFIER_AVX) || \
 	JSONIFIER_CHECK_FOR_INSTRUCTION(JSONIFIER_NEON)
-			if (static_cast<int64_t>(lengthNew) >= 16) {
-				using simd_type						= typename get_type_at_index<simd_internal::avx_list, 0>::type::type;
-				using integer_type					= typename get_type_at_index<simd_internal::avx_list, 0>::type::integer_type;
-				static constexpr int64_t vectorSize = get_type_at_index<simd_internal::avx_list, 0>::type::bytesProcessed;
-				const simd_type search_value		= simd_internal::gatherValue<simd_type>(static_cast<uint8_t>(value));
+			if (lengthNew >= 16) {
+				using simd_type					   = typename get_type_at_index<simd_internal::avx_list, 0>::type::type;
+				using integer_type				   = typename get_type_at_index<simd_internal::avx_list, 0>::type::integer_type;
+				static constexpr size_t vectorSize = get_type_at_index<simd_internal::avx_list, 0>::type::bytesProcessed;
+				const simd_type search_value	   = simd_internal::gatherValue<simd_type>(static_cast<uint8_t>(value));
 				integer_type mask;
 				simd_type chunk;
-				while (static_cast<int64_t>(lengthNew) >= vectorSize) {
+				while (lengthNew >= vectorSize) {
 					chunk = simd_internal::gatherValuesU<simd_type>(data);
 					mask  = static_cast<integer_type>(simd_internal::opCmpEq(chunk, search_value));
 					if JSONIFIER_UNLIKELY (mask != 0) {
@@ -106,13 +106,13 @@ namespace jsonifier_internal {
 			}
 #endif
 			{
-				static constexpr int64_t nBytes{ static_cast<int64_t>(sizeof(uint64_t)) };
-				if (static_cast<int64_t>(lengthNew) >= static_cast<int64_t>(nBytes)) {
+				static constexpr size_t nBytes{ sizeof(size_t) };
+				if (lengthNew >= nBytes) {
 					static constexpr auto valueNew{ repeatByte<value, size_t>() };
 					static constexpr auto highBits{ repeatByte<0x80, size_t>() };
 					static constexpr auto lowBits{ repeatByte<0x01, size_t>() };
 					size_t simdValue;
-					while (static_cast<int64_t>(lengthNew) >= nBytes) {
+					while (lengthNew >= nBytes) {
 						std::memcpy(&simdValue, data, sizeof(size_t));
 						const auto chunk = simdValue ^ valueNew;
 						auto next		 = ((chunk - lowBits) & ~chunk) & highBits;
@@ -128,8 +128,8 @@ namespace jsonifier_internal {
 				}
 			}
 			{
-				static constexpr int64_t nBytes{ static_cast<int64_t>(sizeof(uint32_t)) };
-				if (static_cast<int64_t>(lengthNew) >= static_cast<int64_t>(nBytes)) {
+				static constexpr size_t nBytes{ sizeof(uint32_t) };
+				if (lengthNew >= nBytes) {
 					static constexpr auto valueNew{ repeatByte<value, uint32_t>() };
 					static constexpr auto highBits{ repeatByte<0x80, uint32_t>() };
 					static constexpr auto lowBits{ repeatByte<0x01, uint32_t>() };
@@ -148,8 +148,8 @@ namespace jsonifier_internal {
 				}
 			}
 			{
-				static constexpr int64_t nBytes{ static_cast<int64_t>(sizeof(uint16_t)) };
-				if (static_cast<int64_t>(lengthNew) >= static_cast<int64_t>(nBytes)) {
+				static constexpr size_t nBytes{ sizeof(uint16_t) };
+				if (lengthNew >= nBytes) {
 					static constexpr auto valueNew{ repeatByte<value, uint16_t>() };
 					static constexpr auto highBits{ repeatByte<0x80, uint16_t>() };
 					static constexpr auto lowBits{ repeatByte<0x01, uint16_t>() };
@@ -168,7 +168,7 @@ namespace jsonifier_internal {
 				}
 			}
 
-			if (static_cast<int64_t>(lengthNew) > 0 && *data == value) {
+			if (lengthNew > 0 && *data == value) {
 				++data;
 				return data;
 			}
@@ -179,12 +179,12 @@ namespace jsonifier_internal {
 	struct comparison {
 		template<typename char_type01, typename char_type02> JSONIFIER_FORCE_INLINE static bool compare(const char_type01* lhs, char_type02* rhs, size_t lengthNew) noexcept {
 #if JSONIFIER_CHECK_FOR_INSTRUCTION(JSONIFIER_AVX512)
-			if (static_cast<int64_t>(lengthNew) >= 64) {
-				using simd_type						= typename get_type_at_index<simd_internal::avx_list, 2>::type::type;
-				static constexpr int64_t vectorSize = get_type_at_index<simd_internal::avx_list, 2>::type::bytesProcessed;
-				static constexpr size_t mask		= get_type_at_index<simd_internal::avx_list, 2>::type::mask;
+			if (lengthNew >= 64) {
+				using simd_type					   = typename get_type_at_index<simd_internal::avx_list, 2>::type::type;
+				static constexpr size_t vectorSize = get_type_at_index<simd_internal::avx_list, 2>::type::bytesProcessed;
+				static constexpr size_t mask	   = get_type_at_index<simd_internal::avx_list, 2>::type::mask;
 				simd_type value01, value02;
-				while (static_cast<int64_t>(lengthNew) >= vectorSize) {
+				while (lengthNew >= vectorSize) {
 					value01 = simd_internal::gatherValuesU<simd_type>(lhs);
 					value02 = simd_internal::gatherValuesU<simd_type>(rhs);
 					if (simd_internal::opCmpEq(value01, value02) != mask) {
@@ -197,12 +197,12 @@ namespace jsonifier_internal {
 			}
 #endif
 #if JSONIFIER_CHECK_FOR_INSTRUCTION(JSONIFIER_AVX512) || JSONIFIER_CHECK_FOR_INSTRUCTION(JSONIFIER_AVX2)
-			if (static_cast<int64_t>(lengthNew) >= 32) {
-				using simd_type						= typename get_type_at_index<simd_internal::avx_list, 1>::type::type;
-				static constexpr int64_t vectorSize = get_type_at_index<simd_internal::avx_list, 1>::type::bytesProcessed;
-				static constexpr size_t mask		= get_type_at_index<simd_internal::avx_list, 1>::type::mask;
+			if (lengthNew >= 32) {
+				using simd_type					   = typename get_type_at_index<simd_internal::avx_list, 1>::type::type;
+				static constexpr size_t vectorSize = get_type_at_index<simd_internal::avx_list, 1>::type::bytesProcessed;
+				static constexpr size_t mask	   = get_type_at_index<simd_internal::avx_list, 1>::type::mask;
 				simd_type value01, value02;
-				while (static_cast<int64_t>(lengthNew) >= vectorSize) {
+				while (lengthNew >= vectorSize) {
 					value01 = simd_internal::gatherValuesU<simd_type>(lhs);
 					value02 = simd_internal::gatherValuesU<simd_type>(rhs);
 					if (simd_internal::opCmpEq(value01, value02) != mask) {
@@ -216,12 +216,12 @@ namespace jsonifier_internal {
 #endif
 #if JSONIFIER_CHECK_FOR_INSTRUCTION(JSONIFIER_AVX512) || JSONIFIER_CHECK_FOR_INSTRUCTION(JSONIFIER_AVX2) || JSONIFIER_CHECK_FOR_INSTRUCTION(JSONIFIER_AVX) || \
 	JSONIFIER_CHECK_FOR_INSTRUCTION(JSONIFIER_NEON)
-			if (static_cast<int64_t>(lengthNew) >= 16) {
-				using simd_type						= typename get_type_at_index<simd_internal::avx_list, 0>::type::type;
-				static constexpr int64_t vectorSize = get_type_at_index<simd_internal::avx_list, 0>::type::bytesProcessed;
-				static constexpr size_t mask		= get_type_at_index<simd_internal::avx_list, 0>::type::mask;
+			if (lengthNew >= 16) {
+				using simd_type					   = typename get_type_at_index<simd_internal::avx_list, 0>::type::type;
+				static constexpr size_t vectorSize = get_type_at_index<simd_internal::avx_list, 0>::type::bytesProcessed;
+				static constexpr size_t mask	   = get_type_at_index<simd_internal::avx_list, 0>::type::mask;
 				simd_type value01, value02;
-				while (static_cast<int64_t>(lengthNew) >= vectorSize) {
+				while (lengthNew >= vectorSize) {
 					value01 = simd_internal::gatherValuesU<simd_type>(lhs);
 					value02 = simd_internal::gatherValuesU<simd_type>(rhs);
 					if (simd_internal::opCmpEq(value01, value02) != mask) {
@@ -234,10 +234,10 @@ namespace jsonifier_internal {
 			}
 #endif
 			{
-				static constexpr int64_t nBytes{ static_cast<int64_t>(sizeof(uint64_t)) };
-				if (static_cast<int64_t>(lengthNew) >= static_cast<int64_t>(nBytes)) {
+				static constexpr size_t nBytes{ sizeof(size_t) };
+				if (lengthNew >= nBytes) {
 					size_t v[2];
-					while (static_cast<int64_t>(lengthNew) >= nBytes) {
+					while (lengthNew >= nBytes) {
 						std::memcpy(v, lhs, nBytes);
 						std::memcpy(v + 1, rhs, nBytes);
 						if ((v[0] ^ v[1]) != 0) {
@@ -250,8 +250,8 @@ namespace jsonifier_internal {
 				}
 			}
 			{
-				static constexpr int64_t nBytes{ static_cast<int64_t>(sizeof(uint32_t)) };
-				if (static_cast<int64_t>(lengthNew) >= static_cast<int64_t>(nBytes)) {
+				static constexpr size_t nBytes{ sizeof(uint32_t) };
+				if (lengthNew >= nBytes) {
 					uint32_t v[2];
 					std::memcpy(v, lhs, nBytes);
 					std::memcpy(v + 1, rhs, nBytes);
@@ -264,8 +264,8 @@ namespace jsonifier_internal {
 				}
 			}
 			{
-				static constexpr int64_t nBytes{ static_cast<int64_t>(sizeof(uint16_t)) };
-				if (static_cast<int64_t>(lengthNew) >= static_cast<int64_t>(nBytes)) {
+				static constexpr size_t nBytes{ sizeof(uint16_t) };
+				if (lengthNew >= nBytes) {
 					uint16_t v[2];
 					std::memcpy(v, lhs, nBytes);
 					std::memcpy(v + 1, rhs, nBytes);
